lockfile: reshape for multiplatform
this stubs out for having platform-independent file locking

This still functions the same on `GOOS=linux`, and now succeeds for
`GOOS=windows`. But still fails on `freebsd`, `darwin`, `solaris`, etc
